feat(styleguide): fix JavaScript and font loading issues in Angular v4 demo
Resolved JavaScript error by adding null checks in `app.js` and updating section templates. Fixed Font Awesome font loading by creating a new CSS file with corrected paths and updating build processes. Integrated `section.js` into `section.directive.js` for better maintainability.
