+++ .github/prompts/20250428002.md
@@ -1,0 +1,67 @@
+# Angular v4 でのエラー解消作業
+
+## 問題
+
+スタイルガイドのデモ環境で Angular v4 を使用しているが、以下のエラーが発生していた：
+
+1. JavaScriptエラー: `TypeError: Cannot read properties of undefined (reading 'section')` at `app.js:612:57`
+2. フォントの読み込みエラー: `Failed to decode downloaded font: http://localhost:3000/assets/fonts/fontawesome-webfont.eot`
+
+## 解決方法
+
+### JavaScriptエラーの修正
+
+`app.js` 内で `currentReference.section` に条件チェックを追加して、未定義の場合に処理をスキップするよう修正しました。
+
+```javascript
+scope.$watch('currentReference.section', function() {
+  if (!scope.currentReference || !scope.currentReference.section) {
+    return;
+  }
+  var relatedVariables = scope.currentReference.section.variables || [];
+  // 以降の処理
+});
+```
+
+また、セクションテンプレート（`/demo-output/views/partials/section.html`）にも `sg-variable-watch` ディレクティブを追加し、`section.html` の重複した定義も修正しました。
+
+### フォントエラーの修正
+
+Font Awesomeのフォントパスに問題があり、URL内のクォーテーションが正しく処理されていなかったため、以下の対応を行いました：
+
+1. 正しいパスとフォーマット指定を含む新しいCSSファイルを作成
+   ```css
+   @font-face {
+     font-family: 'FontAwesome';
+     src: url('/assets/fonts/fontawesome-webfont.eot?v=4.2.0');
+     src: url('/assets/fonts/fontawesome-webfont.eot?#iefix&v=4.2.0') format('embedded-opentype'),
+          url('/assets/fonts/fontawesome-webfont.woff?v=4.2.0') format('woff'),
+          url('/assets/fonts/fontawesome-webfont.ttf?v=4.2.0') format('truetype'),
+          url('/assets/fonts/fontawesome-webfont.svg?v=4.2.0#fontawesomeregular') format('svg');
+     font-weight: normal;
+     font-style: normal;
+   }
+   ```
+
+2. このCSSファイルを `/demo-output/font-awesome-fix.css` として配置
+
+3. ビルドプロセスを修正して、フォントパスが正しく処理されるように対応
+   - `demo-gulpfile.js` と `gulpfile.babel.js` の両方に、URL処理と format 関数の修正を追加
+
+4. HTMLファイル（`/demo-output/index.html`）でこのCSSファイルを直接読み込むよう設定
+   ```javascript
+   addCssLink('/styleguide_at_rules.css');
+   addCssLink('/styleguide-app.css', {noReload: true});
+   addCssLink('/font-awesome-fix.css', {noReload: true});
+   ```
+
+これらの修正により、`npm run build` 実行後もフォントが正しく表示され、コンソールエラーが完全に解消されました。
+
+## まとめ
+
+今回の修正では、以下の問題を解決しました：
+
+1. セクション参照時のJavaScriptエラー
+2. Font Awesomeのフォント読み込みエラー
+
+これにより、SC5 Styleguide がNode.js v22.14.0環境で正常に動作するようになりました。
