+++ lib/app/css/font-awesome-fix.css
@@ -1,0 +1,14 @@
+/**
+ * Special fix for Font Awesome font paths
+ * This file will be loaded directly in the HTML to ensure the fonts are correctly loaded
+ */
+@font-face {
+  font-family: 'FontAwesome';
+  src: url('/assets/fonts/fontawesome-webfont.eot?v=4.2.0');
+  src: url('/assets/fonts/fontawesome-webfont.eot?#iefix&v=4.2.0') format('embedded-opentype'),
+       url('/assets/fonts/fontawesome-webfont.woff?v=4.2.0') format('woff'),
+       url('/assets/fonts/fontawesome-webfont.ttf?v=4.2.0') format('truetype'),
+       url('/assets/fonts/fontawesome-webfont.svg?v=4.2.0#fontawesomeregular') format('svg');
+  font-weight: normal;
+  font-style: normal;
+}
