+++ lib/app/index.html
@@ -28,6 +28,7 @@
 
       addCssLink('{{{appRoot}}}/styleguide_at_rules.css');
       addCssLink('{{{appRoot}}}/styleguide-app.css', {noReload: true});
+      addCssLink('{{{appRoot}}}/css/components/font-awesome/font-awesome-fix.css', {noReload: true});
 
       // Use Shadow DOM but fallback to global styles if shadow DOM is not supported
       if (typeof head.createShadowRoot !== 'function' || _styleguideConfig.disableEncapsulation) {
