+++ lib/app/js/directives/design.js
@@ -31,6 +31,9 @@ angular.module('sgApp')
         scope.showRelated = true;
 
         scope.$watch('currentReference.section', function() {
+          if (!scope.currentReference || !scope.currentReference.section) {
+            return;
+          }
           var relatedVariables = scope.currentReference.section.variables || [];
           if (scope.showRelated && relatedVariables.length === 0 && scope.sections.data) {
             parentRef = scope.currentReference.section.reference;
