--- lib/app/views/partials/section.html
@@ -1,91 +1,0 @@
-<section id="section-{{section.reference}}" class="sg sg-section"
-  ng-class="{active: status.isActive, main: section.isEmptyMainSection}">
-  <header class="sg sg-section-header">
-    <a ui-sref="app.index.section({section: '{{section.reference}}'})">
-      <i class="fa fa-link"></i>
-      <h1 class="sg">
-        <span class="sg-reference-number" ng-if="$root.config.data.showReferenceNumbers">{{ section.reference }}</span>
-        <span ng-bind-html="section.header | unsafe"></span>ction-{{section.reference}}" class="sg sg-section"
-  ng-class="{active: section.isActive, main: section.isEmptyMainSection}">
-  <header class="sg sg-section-header">
-    <a ui-sref="app.index.section({section: '{{section.reference}}'})">
-      <i class="fa fa-link"></i>
-      <h1 class="sg">
-        <span class="sg-reference-number" ng-if="$root.config.data.showReferenceNumbers">{{ section.reference }}</span>
-        <span ng-bind-html="section.header | unsafe"></span> id="section-{{section.reference}}" class="sg sg-section"
-  ng-class="{active: isActive(section), main: isEmptyMainSection(section)}">
-  <header class="sg sg-section-header">
-    <a ui-sref="app.index.section({section: '{{section.reference}}'})">
-      <i class="fa fa-link"></i>
-      <h1 class="sg">
-        <span class="sg-reference-number"  ng-if="config.data.showReferenceNumbers">{{ section.reference }}</span>
-        <span ng-bind-html="section.header | unsafe"></span>
-        <span class="sg-section-source-ref" ng-if="!config.data.showReferenceNumbers">{{ section.reference }}</span>
-        <span class="sg-section-source">{{ section.file }} </span>
-      </h1>
-    </a>
-  </header>
-  <div class="sg sg-section-partial" ng-if="section.modifiers.length || section.description">
-    <div class="sg sg-section-description" ng-bind-html="section.description | unsafe"></div>
-    <ul ng-if="section.modifiers.length" class="sg modifier-list">
-      <li class="sg-item" ng-repeat="mod in section.modifiers">
-        <strong ng-bind-html="mod.name | unsafe"></strong> - <span ng-bind-html="mod.description | unsafe"></span>
-      </li>
-    </ul>
-  </div>
-  <div ng-repeat-start="mod in section.modifiers"></div>
-  <div class="sg sg-section-partial">
-    <div class="sg sg-label">
-      <a ui-sref="app.fullscreen({section: '{{section.reference}}-{{mod.id}}'})" target="_blank">
-        <span class="sg sg-name" ng-bind-html="mod.name | unsafe"></span><i class="fa fa-arrows-alt"></i>
-      </a>
-    </div>
-    <shadow-dom id="section-{{section.reference}}-modifier-{{mod.id}}-render-markup">
-      <div class="section-markup-wrapper" ng-bind-html="mod.renderMarkup | setModifierClass: mod.className | setVariables: variables | addWrapper | unsafe" dynamic-compile></div>
-    </shadow-dom>
-  </div>
-  <div ng-repeat-end></div>
-  <div
-    class="sg sg-section-partial"
-    ng-if="!section.modifiers.length && section.markup">
-    <div class="sg sg-label">
-      <a ui-sref="app.fullscreen({section: '{{section.reference}}'})" target="_blank">
-        <i class="fa fa-arrows-alt"></i>
-      </a>
-    </div>
-
-    <shadow-dom id="section-{{section.reference}}-render-markup">
-      <div class="section-markup-wrapper" ng-bind-html="section.renderMarkup | setVariables: variables | addWrapper | unsafe" dynamic-compile></div>
-    </shadow-dom>
-  </div>
-
-  <div class="sg sg-section-partial sg-code-listing" ng-if="section.css">
-    <div ng-if="section.showCSS" class="sg sg-label">
-      <a ng-click="section.showCSS = false" target="_blank">
-        <i class="fa fa-close"></i>
-      </a>
-    </div>
-    <pre class="sg" ng-if="section.showCSS"><code hljs hljs-language="css" hljs-source="section.css" class="css"></code></pre>
-    <a class="sg sg-show-section" ng-hide="section.showCSS" ng-click="section.showCSS = true">Show CSS</a>
-  </div>
-
-  <div class="sg sg-section-partial sg-code-listing" ng-if="section.markupJade">
-    <div ng-if="section.showMarkupJade" class="sg sg-label">
-      <a ng-click="section.showMarkupJade = false" target="_blank">
-        <i class="fa fa-close"></i>
-      </a>
-    </div>
-    <pre class="sg" ng-if="section.showMarkupJade"><code hljs hljs-source="section.markupJade" class="jade"></code></pre>
-    <a class="sg sg-show-section" ng-hide="section.showMarkupJade" ng-click="section.showMarkupJade = true">Show JADE markup</a>
-  </div>
-
-  <div class="sg sg-section-partial sg-code-listing" ng-if="section.markup">
-    <div ng-if="section.showMarkup" class="sg sg-label">
-      <a ng-click="section.showMarkup = false" target="_blank">
-        <i class="fa fa-close"></i>
-      </a>
-    </div>
-    <pre class="sg" ng-if="section.showMarkup"><code hljs hljs-language="html" hljs-source="section.markup" class="html"></code></pre>
-    <a class="sg sg-show-section" ng-hide="section.showMarkup" ng-click="section.showMarkup = true">Show markup</a>
-  </div>
-</section>
